Add SetStackSize to usercorn and use in DOS loader
Since the COM loader uses NewUsercornRaw and maps everything in manually
it counts on the DOS kernel to setup some of the rest of the state.

`-v` was crashing when trying to print a stack trace because the stack
size wasn't correct set. To set this I needed to expose SetStackSize.